Clean up code and remove some duplication
