+++ packages/arbor-ost/src/handlers/$set/visitors/entries.ts
@@ -1,16 +1,12 @@
 import { Visitor } from "../../visitor"
-import { isProxiable } from "../../$object/visitors/proxiable"
 
 export class EntriesVisitor extends Visitor {
   prop = "entries"
 
-  visit({ target, ost, $node }) {
+  visit({ $node }) {
     return function* () {
-      for (const value of target.values()) {
-        const nodeValue = isProxiable(value)
-          ? ost.nodeOf(value) || $node.$createChild(value)
-          : value
-        yield [nodeValue, nodeValue]
+      for (const child of $node[Symbol.iterator]()) {
+        yield [child, child]
       }
     }
   }
