+++ packages/arbor-ost/src/handlers/$set/visitors/values.ts
@@ -1,17 +1,12 @@
 import { Visitor } from "../../visitor"
-import { isProxiable } from "../../$object/visitors/proxiable"
 
 export class ValuesVisitor extends Visitor {
   prop = "values"
 
-  visit({ target, ost, $node }) {
+  visit({ $node }) {
     return function* () {
-      for (const value of target.values()) {
-        if (isProxiable(value)) {
-          yield ost.nodeOf(value) || $node.$createChild(value)
-        } else {
-          yield value
-        }
+      for (const child of $node[Symbol.iterator]()) {
+        yield child
       }
     }
   }
